+++ tests/testthat/test-mod_StudyOverview.R
@@ -22,14 +22,14 @@ test_that("mod_StudyOverview_Server sets selected group and metric", {
       dfGroups = sample_dfGroups,
       dfBounds = sample_dfBounds,
       rctv_strMetricID = reactiveVal("Analysis_kri0002"),
-      rctv_strGroupID = reactiveVal("0X003")
+      rctv_strGroupID = reactiveVal("0X2096")
     ),
     {
       # Test the selected group reactive
       session$setInputs(`scatter-Analysis_kri0001-plot` = "All")
       expect_equal(rctv_strGroupID(), "All")
-      session$setInputs(`scatter-Analysis_kri0001-plot` = "0X005")
-      expect_equal(rctv_strGroupID(), "0X005")
+      session$setInputs(`scatter-Analysis_kri0001-plot` = "0X5985")
+      expect_equal(rctv_strGroupID(), "0X5985")
 
       # Test the selected metric reactive
       # When we click a plot, it updates.