+++ NEWS.md
@@ -8,6 +8,7 @@ We now support multiple group levels! We also provided more options to allow app
 - We overhauled the way data flows into plugins, allowing for greater flexibility in how plugins function. Importantly, plugins now have access to Subject-, Group-, *and* Study-level domain data. Plugins can also access the gsm Data Model data (`dfAnalyticsInput`, `dfBounds`, `dfGroups`, `dfMetrics`, and `dfResults`). As part of these changes, data fetching has been optimized to subset cached data when available. We hope this will speed up some slow operations (#430, #435, #438, #440, #447, #450).
 
 ## Other improvements
+- The sample data is now more robust, and has already allowed us to find one issue which might turn up in real data (when a particular site doesn't have data for one or more KRIs) (#398).
 - Each tab now has more control over its own interface, rather than requiring a large sidebar on every tab. As a result, the "Metric Details" tab has a more streamlined (and hopefully easier to navigate) interface (#428).
 - `dfAnalyticsInput` can now provide information from multiple `SnapshotDate`s (#455).